Do not build for OSX
the OSX build has some problems completing its test suite. Upstream is working
on it at https://github.com/BRiAl/BRiAl/pull/25 but it is not entirely clear
whether it's sane just to skip those tests or whether something is actually
broken.

I am not sure what's the policy here but it is maybe best to ship 1.2.1 only
for Linux where it's known to work and release it for OSX once upstream has
fixed it.